Created center_eval function
Created the center_eval function which checks a given string for php
code, extracts the lines of code and executes it (anything in between
<?php ?> tags), then replaces the code snippet in the string with the
output of the executed code. Doing this instead of just using straight
up eval allows me to be selective about which functions are allowed to
be executed and to separate the php tags from the text.